Revert "ValueFlow: extracted `ValueFlowAnalyzer` into separate file (#6640)"
This reverts commit d7eb66d85d3ff484461db848df548da79f21341c.
