+++ lib/cppcheck.vcxproj
@@ -87,7 +87,6 @@
     <ClCompile Include="token.cpp" />
     <ClCompile Include="tokenlist.cpp" />
     <ClCompile Include="utils.cpp" />
-    <ClCompile Include="valueflowanalyzer.cpp" />
     <ClCompile Include="vf_array.cpp" />
     <ClCompile Include="vf_arraybool.cpp" />
     <ClCompile Include="vf_arrayelement.cpp" />
@@ -187,7 +186,6 @@
     <ClInclude Include="tokenrange.h" />
     <ClInclude Include="utils.h" />
     <ClInclude Include="valueflow.h" />
-    <ClInclude Include="valueflowanalyzer.h" />
     <ClInclude Include="valueptr.h" />
     <ClInclude Include="version.h" />
     <ClInclude Include="vf_analyze.h" />
