+++ lib/lib.pri
@@ -75,7 +75,6 @@ HEADERS += $${PWD}/addoninfo.h \
            $${PWD}/tokenrange.h \
            $${PWD}/utils.h \
            $${PWD}/valueflow.h \
-           $${PWD}/valueflowanalyzer.h \
            $${PWD}/valueptr.h \
            $${PWD}/version.h \
            $${PWD}/vf_analyze.h \
@@ -166,7 +165,6 @@ SOURCES += $${PWD}/valueflow.cpp \
            $${PWD}/token.cpp \
            $${PWD}/tokenlist.cpp \
            $${PWD}/utils.cpp \
-           $${PWD}/valueflowanalyzer.cpp \
            $${PWD}/vf_array.cpp \
            $${PWD}/vf_arraybool.cpp \
            $${PWD}/vf_arrayelement.cpp \
